requirements.txt
aiogram==3.15.0
aiosqlite==0.20.0
gspread==6.1.2
google-auth==2.37.0
apscheduler==3.10.4
python-dotenv==1.0.1
pytz==2024.1
